third_party/blink/web_tests/platform/generic/svg/custom/frame-getSVGDocument-expected.txt

Tests that HTMLIFrameElement, HTMLEmbedElement and HTMLObjectElement expose a getSVGDocument accessor

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS !!iframe.getSVGDocument is true
PASS iframe.getSVGDocument().toString() is "[object XMLDocument]"
PASS !!embed.getSVGDocument is true
PASS embed.getSVGDocument().toString() is "[object XMLDocument]"
PASS !!object.getSVGDocument is true
PASS object.getSVGDocument().toString() is "[object XMLDocument]"
PASS successfullyParsed is true

TEST COMPLETE
